Accept floating-point polling interval

+++ lib/synapse/service_watcher/zookeeper_poll.rb
@@ -57,7 +57,7 @@ class Synapse::ServiceWatcher
       raise ArgumentError, "zookeeper poll watcher expects zookeeper_poll method" unless @discovery['method'] == 'zookeeper_poll'
       raise ArgumentError, "zookeeper poll watcher expects integer polling_interval_sec >= 0" unless (
           @discovery.has_key?('polling_interval_sec') &&
-          @discovery['polling_interval_sec'].is_a?(Integer) &&
+          @discovery['polling_interval_sec'].is_a?(Numeric) &&
           @discovery['polling_interval_sec'] >= 0
         )
       raise ArgumentError, "missing or invalid zookeeper host for service #{@name}" \
